gsm/ipa: fix t_len truncation in ipa_ccm_id_resp_parse()
The IPA CCM ID RESP TLV format uses a 16bit length field, and
osmo_load16be() is used to read it, but t_len was declared as
uint8_t, silently truncating any length above 255.

Related: OS#7050

+++ src/gsm/ipa.c
@@ -209,7 +209,7 @@ int ipa_ccm_id_get_parse(struct tlv_parsed *dec, const uint8_t *buf, unsigned in
  *  \returns 0 on success; negative on error */
 int ipa_ccm_id_resp_parse(struct tlv_parsed *dec, const uint8_t *buf, unsigned int len)
 {
-	uint8_t t_len;
+	uint16_t t_len;
 	uint8_t t_tag;
 	const uint8_t *cur = buf;
